+++ client/src/components/PostDetails/PostDetails.tsx
@@ -43,6 +43,8 @@ export default function PostDetails() {
 
   async function deleteComment(commentId: string) {
     if (!post) return
+    let confirm = window.confirm("Do you really want to delete this comment?")
+    if (!confirm) return
     await apiDeleteComment(post._id, commentId)
     setComments((prev) => prev.filter((comment) => comment._id !== commentId))
     setPost(prev => prev ? { ...prev, commentAmount: prev.commentAmount - 1 } : prev)
@@ -59,6 +61,8 @@ export default function PostDetails() {
   }
 
   const onDeletePost = async (postId: string) => {
+    let confirm = window.confirm("Do you really want to delete this post?")
+    if (!confirm) return
     await deletePost(postId)
     navigate(-1)
   }
